Begin mongo transition, some refactoring
