+++ package.json
@@ -26,11 +26,11 @@
     "cookie-parser": "1.3.2",
     "ejs": "1.0.0",
     "express": "4.9.4",
-    "express-mysql-session": "*",
+    "connect-mongo": "^0.8.2",
     "express-session": "*",
     "minimist": "^1.1.1",
     "moment": "^2.10.3",
-    "mysql": "*",
+    "mongodb": "~2.0",
     "nconf": "*",
     "node-twitter": "*",
     "p3p": "0.0.2",
